# openapi.yaml
openapi: 3.1.0
info:
  title: GPT Legal Forensic API
  description: API para análisis legal automatizado con múltiples módulos...
  version: "1.0"
servers:
  - url: https://gpt-legal-backend.onrender.com
    description: Servidor público en Render
paths:
...
  /bcn/api/bcn/norma/{id_norma}:
    get:
      operationId: obtenerNormaBcn
      summary: Obtener norma legal por ID desde BCN
      parameters:
        - name: id_norma
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Norma legal obtenida
          content:
            application/json:
              schema:
                type: object
                properties:
                  norma:
                    type: object
                    additionalProperties: true
  /legislacion/buscar:
    get:
      operationId: buscarNorma
      summary: Buscar norma legal en BCN por palabra clave
      parameters:
        - name: palabra
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Resultados de búsqueda
          content:
            application/json:
              schema:
                type: object
                properties:
                  resultados:
                    type: array
                    items:
                      type: object
                      additionalProperties: true
  /legislacion/norma/{id_norma}:
    get:
      operationId: obtenerNormaLegislacion
      summary: Obtener norma desde legislacion_bcn por ID
      parameters:
        - name: id_norma
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Norma encontrada
          content:
            application/json:
              schema:
                type: object
                properties:
                  norma:
                    type: object
                    additionalProperties: true
...